Created a Fritzing diagram for the hardware
The hardware used in this project is open sourced in the form of this Fritzing project file. Note that the HC-SR04 sensors are there mainly for visual purposes, I am using 2 HY-SRF05 sensors and are facing away from the wiring. Also the resistors are not exactly correct. The ratio is 2:1 to get the Voltage from 5hz to 3.33Hz.